reformated history recording; added genealogy tracking; division rate is now dependent on population density

+++ simulations/scripts/ChemostatSimulation.py
@@ -17,9 +17,9 @@ import numpy as np
 
 
 class Chemostat:
-    def __init__(self, volume_val: float, dilution_rate_val: float, n_cells=None):
+    def __init__(self, volume_val: float, dilution_rate: float, n_cells=None):
         self.V = volume_val
-        self.dilution_rate = dilution_rate_val
+        self.D = dilution_rate
         self.cells = []
         if n_cells:
             self.populate_with_cells(n_cells)
@@ -29,10 +29,10 @@ class Chemostat:
         return len(self.cells)
 
     def populate_with_cells(self, n_cells: int) -> None:
-        self.cells += [Cell(i) for i in range(n_cells)]
+        self.cells += [Cell(chemostat=self, cell_id=i) for i in range(n_cells)]
 
     def dilute(self) -> None:
-        expected_n_cells_to_remove = self.dilution_rate * (self.N**2)/self.V
+        expected_n_cells_to_remove = self.D * self.N/self.V
         n_cells_to_remove = np.random.poisson(expected_n_cells_to_remove, 1)[0]
         dead_cells = np.random.choice(self.cells, size=min(self.N, n_cells_to_remove), replace=False)
         for cell in dead_cells:
@@ -43,8 +43,14 @@ class Cell:
     critical_amount = 10
     rate_of_accumulation = 1
 
-    def __init__(self, cell_id: int, age=0):
+    def __init__(self,
+                 chemostat: Chemostat,
+                 cell_id: int,
+                 parent_id=None,
+                 age=0):
+        self.chemostat = chemostat
         self.id = cell_id
+        self.parent_id = parent_id
         self.age = age
         self.has_reproduced = False
         self.has_died = False
@@ -56,14 +62,15 @@ class Cell:
         at a given age
         """
         alpha = Cell.critical_amount + 1
-        beta = Cell.rate_of_accumulation
+        beta = Cell.rate_of_accumulation/self.chemostat.N
         return gamma.pdf(self.age, a=alpha, scale=1/beta) / max(1 - gamma.cdf(self.age, a=alpha, scale=1/beta),
                                                                 gamma.pdf(self.age, a=alpha, scale=1/beta))
 
     def reproduce(self, offspring_id: int) -> list:
         self.has_reproduced = np.random.uniform(0, 1) < self.prob_of_division
         if self.has_reproduced:
-            return [Cell(self.id), Cell(offspring_id)]
+            return [Cell(chemostat=self.chemostat, cell_id=offspring_id, parent_id=self.id),
+                    Cell(chemostat=self.chemostat, cell_id=offspring_id+1, parent_id=self.id)]
         else:
             return [self]
 
@@ -72,24 +79,16 @@ class Cell:
 
 
 class Simulation:
-    def __init__(self, chemostat_obj: Chemostat, carrying_capacity_val: int):
+    def __init__(self, chemostat_obj: Chemostat, save_path: str):
         self.chemostat = chemostat_obj
-        self.density_coefficient = self._calculate_density_coefficient(carrying_capacity_val,
-                                                                       self.chemostat.dilution_rate,
-                                                                       self.chemostat.V)
-        self.history = History(self)
+        self.history = History(self, save_path)
 
     @property
     def current_max_id(self) -> int:
         return max([cell.id for cell in self.chemostat.cells])
 
-    @staticmethod
-    def _calculate_density_coefficient(carrying_capacity_val: int, dilution_rate_val: float, volume_val: float) -> float:
-        return carrying_capacity_val/(np.log(1/2)/(2*np.log(1-dilution_rate_val/volume_val)))
-
-    def step(self) -> None:
+    def step(self, step_number: int) -> None:
         # Cells are diluted
-        # for _ in range(int(round(self.chemostat.N/self.density_coefficient))):
         self.chemostat.dilute()
 
         # Alive cells reproduce
@@ -98,9 +97,9 @@ class Simulation:
             new_cells += cell.reproduce(self.current_max_id+1)
 
         # History is recorded
-        self.history.record()
+        self.history.record(step_number)
 
-        # Move to the next generation
+        # Move to the next time step
         self.chemostat.cells = new_cells
 
         # Time passes
@@ -108,27 +107,38 @@ class Simulation:
             cell.age += 1
 
     def run(self, n_steps: int) -> None:
-        for _ in range(n_steps):
-            self.step()
+        for step_number in range(n_steps):
+            self.step(step_number)
+        self.history.save()
 
 
 class History:
-    def __init__(self, simulation_obj: Simulation):
+    def __init__(self, simulation_obj: Simulation, save_path: str):
         self.simulation = simulation_obj
-        self.history_table = pd.DataFrame(columns=["generation", "n_cells"])
-        self.cells_table = pd.DataFrame(columns=["generation", "cell_id", "cell_age",
-                                                 "cell_damage", "has_divided", "has_died"])
 
-    def record(self) -> None:
+        self.save_path = save_path
+        self.run_id = round(datetime.datetime.now().timestamp())
+        (Path(self.save_path) / Path(str(self.run_id))).mkdir()
+
+        self.history_table, self.cells_table, self.genealogy_table = None, None, None
+        self.reset()
+
+    def reset(self):
+        self.history_table = pd.DataFrame(columns=["time_step", "n_cells"])
+        self.cells_table = pd.DataFrame(columns=["time_step", "cell_id", "cell_age",
+                                                 "cell_damage", "has_divided", "has_died"])
+        self.genealogy_table = pd.DataFrame(columns=["cell_id", "parent_id"])
+
+    def record(self, time_step) -> None:
         # Make a record history_table
         row = pd.DataFrame.from_dict(
-            {"generation": [len(self.history_table) + 1], "n_cells": [self.simulation.chemostat.N]})
+            {"time_step": [time_step], "n_cells": [self.simulation.chemostat.N]})
         self.history_table = pd.concat([self.history_table, row], ignore_index=True)
 
         # Make a record in cells_table
         cells = self.simulation.chemostat.cells
         df_to_add = pd.DataFrame.from_dict(
-            {"generation": [len(self.history_table) + 1 for _ in range(len(cells))],
+            {"time_step": [time_step for _ in range(len(cells))],
              "cell_id": [cell.id for cell in cells],
              "cell_age": [cell.age for cell in cells],
              "has_divided": [cell.has_reproduced for cell in cells],
@@ -137,25 +147,35 @@ class History:
 
         self.cells_table = pd.concat([self.cells_table, df_to_add], ignore_index=True)
 
-    def save(self, output_folder: str) -> None:
-        run_id = round(datetime.datetime.now().timestamp() % 1000000000)
-        subfolder = Path(output_folder)/Path(str(run_id))
-        subfolder.mkdir()
-        history_file_path = subfolder/Path(f"history_{run_id}.tsv")
-        cells_file_path = subfolder/Path(f"cells_{run_id}.tsv")
-        for table, path in zip([self.history_table, self.cells_table],
-                               [history_file_path, cells_file_path]):
-            table.to_csv(path, sep="\t", index=False)
+        # Make a record in genealogy_table
+        cells = list(filter(lambda el: el.id not in self.genealogy_table.cell_id, self.simulation.chemostat.cells))
+        df_to_add = pd.DataFrame.from_dict(
+            {"cell_id": [cell.id for cell in cells],
+             "parent_id": [cell.parent_id for cell in cells],
+             })
+        self.genealogy_table = pd.concat([self.genealogy_table, df_to_add], ignore_index=True)
+
+        if len(self.cells_table) > 10000:
+            self.save()
+
+    def save(self) -> None:
+        for table, stem in zip([self.history_table, self.cells_table, self.genealogy_table],
+                               ["history", "cells", "genealogy"]):
+            path = Path(f"{self.save_path}/{self.run_id}/{stem}_{self.run_id}.tsv")
+            if path.exists():
+                table.to_csv(path, mode="a", sep="\t", index=False, header=False)
+            else:
+                table.to_csv(path, sep="\t", index=False)
+        self.reset()
 
     def __str__(self):
         return str(self.history_table)
 
 
 if __name__ == "__main__":
-    volume, dilution_rate, carrying_capacity = map(int, sys.argv[1:4])
+    volume, dilution_rate = map(int, sys.argv[1:3])
     simulation = Simulation(Chemostat(volume_val=volume,
-                                      dilution_rate_val=dilution_rate,
+                                      dilution_rate=dilution_rate,
                                       n_cells=1),
-                            carrying_capacity_val=carrying_capacity)
-    simulation.run(500)
-    simulation.history.save("../data/")
+                            save_path="../data/")
+    simulation.run(10000)
